Update session.middleware.js
Merging session-id.middleware.js to
session .middleware.js

 adding chalk , config and actions to   requiremenrs

 adding variable is post

we then  using router.use we add code from session-id  in lines 13 to 16 and in line 17 we add a  17  else condition. on the code from

Merging complete thanks for letting me contribute to your project.

Im not deleting your session-id as I believe the code is needed elsewhere and   I would need more knowledge to do this.

have a great day